Adds SECURITY, load test
The load test demonstrates that the search API falls over at 256
simultaneous connections. This says that I am not releasing *something*.
Probably HTTP responses are not closing.

Rethinking the API (again) for simplicity is probably a good idea. the
underlying structures/architecture is over-complicated. It would be easy
to clean up, and probably get to a space where thousands of simultaneous
users are possible.

(Because the underlying databases are RO, there is no concurrency limit
for queries... and Go should be able to scale better. Hence, something
is wrong.)