With E2E coming together
This adds the tooling to crawl a local site at test.gov.

What is odd is that the SQLite databases are *slightly* different sizes,
but they have the same number of pages.

I don't think there is anything non-deterministic in this... well, the
workers are. But, the fact that I get 2721 from three crawls (one alone
on t1.gov, and one together on t2.gov and t3.gov) suggests to me that
we're reasonably robust for continuing to experiment.