Adding the admin/admin CLI
This will be how we control the app, at least until we have more
people/engineers.

The `admin` component will let us do an OAuth dance, and the admin-cli
will give a modest interface to the things we can do. It will talk via
API to the `admin` component, and... from there, we talk to the queue.