Pack updates and Entree
Pack now only creates a DB at the end. Instead of opening one on first
receipt of a `host`, it just tracks the last extracted time, and when we
time out, the DB is opened, written, and closed in one swell foop.

Entree is a new service, and needs a doc file. That's next.