A migrator for jemison-db
In theory, we can run `dbmate`.

However, that involves parsing VCAP_SERVICES in bash or some other
scripting language, just so we can run `dbmate`. The alternative was to
yank a few lines out of `entree` and run this as a standalone app. We
can integrate it as a single executable into any of our services, and
run it there. It is less fragile than bash-fu.