This is a breaking commit
This will break `pack`.

However, I wanted to commit what I had.

This introduces a new content model for the search databases. It breaks
out URLs, titles, headers, and the body content.

This will let us compose a query against the SQLite databases that
provides some ranking based on what the content is. E.g. we can rank a
title higher than the body content. (And, if we want to, we can rank an
H1 higher than an H2, and higher than a paragraph.)

That level of nuance will take some fiddling with the SQL for the query,
but the nice thing is that it is testable. We might even be able to
offer to have different query ranking algorithms for different tables.
This would then become a matter of defining those queries once, in the
model, and then select them based on the domain.

Or, not.

For now, though, this puts the initial model in, and sets me up for some
continued implementation on an improved data model later.