Still haven't found...
`sqlc` cannot handle FTS5 tables in all their forms. Annoying.

So, a custom `Search()` had to be written. This... could be worse. It is
parameterized.

Tests are in place. The search function is doing what it should; titles
are prioritized over headers, which is prioritized over body content.

Not sure yet about returning snippets. But, that can be next. Might have
to be a second query. Mostly, my SQL isn't what it could be, and perhaps
there's a way.

Still "in-progress." Services will break from this commit.

Next time, I'll use a branch instead of breaking `main`.