Moving the Search() method
This puts it in with the other queries, and makes it look the same.

This involves putting it in with the generated functions from sqlc, but
it makes more sense here. From a whole-project import perspective, it
makes code more consistent. And, it lets the query accsess the DB, etc.