Simplifies full crawl logic
I'll update diagrams in the docs.

This simplifies things by moving the next_fetch to the `hosts` table.

Instead of tracking when things were fetched per-page, we fetch
per-host.

When we issue a pass, it is either for a single page, or for a full
host. A full host resets the next_fetch date, allowing the crawl to
proceed.