Fixing connection count issues?
It may be that the init for queues was breaking because too many
connections were being made (e.g. one or more connections in every
worker) instead of a single pool for all of the workers.

This has to do with how I configured the `InsertX` functions. They
basically each develop a pool. As a result, scaling `fetch` to 100
workers created 100+ connections, which exceeds the Postgres default.

There's no need for that kind of silliness, so instead, it will get
cleaned up. Testing.