Cleanup of queueing
I was doing a lot of weird things around the queueing.

Now, there's one piece of code handling all of it, and all of the
servicse use it. This is cleaner, and now any issues can be addressed in
one place.

Found a bug in the S3 code... when we write to Minio too fast, it throws
errors. This will (probably?) not matter in production, but the handling
of the error gracefully is an improvement, and probably will eliminate
the failed `walk` jobs.

Nginx is now used as a balancer in front of two Minio instances in all
configs.