Fixed serve, TF as JSON
This is an experiment in TF as JSON.

https://developer.hashicorp.com/terraform/language/syntax/json#block-type-specific-exceptions

Specifically, I'm using Jsonnet to generate TF config as JSON, which
eliminates some of the syntactic pieces of TF config as a hurdle, and
eliminates a lot of copy-paste. I can now use variables, and *generate*
the TF exactly the way I want.

(Yes, TF has variables.)

But, I can build a composable config with functions, variables,
comprehensions... and, it get syntax checked before going to JSON.

This version validates. I don't know if it deploys, because I'm out of
quota.