Inheritance for partitioning
It turns out that ranges... well, actually, they might still work.

I've got scripts for both. The issue is handling things we don't have a
range for, which... perhaps shouldn't happen.

Anyway. This is an evolving experiment.

This uses inheritance tables. It still relies on the domain64 value in
the trigger function to route things to the correct table.

To be seen, with some actual performance testing.